+++ Sources/Sentry/SentryAsyncSafeLog.h
@@ -47,8 +47,7 @@ extern "C" {
 
 static char g_logFilename[1024];
 
-void sentry_asyncLogC(
-    const char *level, const char *file, int line, const char *function, const char *fmt, ...);
+void sentry_asyncLogC(const char *level, const char *file, int line, const char *fmt, ...);
 
 #define i_SENTRY_ASYNC_SAFE_LOG sentry_asyncLogC
 
@@ -62,7 +61,7 @@ void sentry_asyncLogC(
 #define SENTRY_ASYNC_SAFE_LOG_LEVEL SENTRY_ASYNC_SAFE_LOG_LEVEL_ERROR
 
 #define a_SENTRY_ASYNC_SAFE_LOG(LEVEL, FMT, ...)                                                   \
-    i_SENTRY_ASYNC_SAFE_LOG(LEVEL, __FILE__, __LINE__, __PRETTY_FUNCTION__, FMT, ##__VA_ARGS__)
+    i_SENTRY_ASYNC_SAFE_LOG(LEVEL, __FILE__, __LINE__, FMT, ##__VA_ARGS__)
 
 // ============================================================================
 #pragma mark - API -