+++ Sources/Sentry/include/HybridPublic/SentryAppStartMeasurement.h
@@ -17,7 +17,7 @@ SENTRY_NO_INIT
 @end
 
 /**
- * @warning This feature is not available in @c Debug_without_UIKit and @c Release_without_UIKit
+ * @warning This feature is not available in @c DebugWithoutUIKit and @c ReleaseWithoutUIKit
  * configurations even when targeting iOS or tvOS platforms.
  */
 @interface SentryAppStartMeasurement : NSObject