+++ Tests/SentryTests/Integrations/SessionReplay/SentrySessionReplayTests.swift
@@ -13,34 +13,37 @@ class SentrySessionReplayTests: XCTestCase {
     }
      
     private class TestReplayMaker: NSObject, SentryReplayVideoMaker {
-        
         var videoWidth: Int = 0
         var videoHeight: Int = 0
+         
+        var screens = [String]()
         
         struct CreateVideoCall {
-            var duration: TimeInterval
             var beginning: Date
+            var end: Date
             var outputFileURL: URL
             var completion: ((Sentry.SentryVideoInfo?, Error?) -> Void)
         }
         
         var lastCallToCreateVideo: CreateVideoCall?
-        func createVideoWith(duration: TimeInterval, beginning: Date, outputFileURL: URL, completion: @escaping (Sentry.SentryVideoInfo?, (Error)?) -> Void) throws {
-            lastCallToCreateVideo = CreateVideoCall(duration: duration,
-                                                        beginning: beginning,
-                                                        outputFileURL: outputFileURL,
-                                                        completion: completion)
+        func createVideoWith(beginning: Date, end: Date, outputFileURL: URL, completion: @escaping (Sentry.SentryVideoInfo?, (Error)?) -> Void) throws {
+            lastCallToCreateVideo = CreateVideoCall(beginning: beginning,
+                                                    end: end,
+                                                    outputFileURL: outputFileURL,
+                                                    completion: completion)
             
             try? "Video Data".write(to: outputFileURL, atomically: true, encoding: .utf8)
             
-            let videoInfo = SentryVideoInfo(path: outputFileURL, height: 1_024, width: 480, duration: duration, frameCount: 5, frameRate: 1, start: beginning, end: beginning.addingTimeInterval(duration), fileSize: 10)
+            let videoInfo = SentryVideoInfo(path: outputFileURL, height: 1_024, width: 480, duration: end.timeIntervalSince(beginning), frameCount: 5, frameRate: 1, start: beginning, end: end, fileSize: 10, screens: screens)
             
             completion(videoInfo, nil)
         }
         
         var lastFrame: UIImage?
-        func addFrameAsync(image: UIImage) {
+        func addFrameAsync(image: UIImage, forScreen: String?) {
             lastFrame = image
+            guard let forScreen = forScreen else { return }
+            screens.append(forScreen)
         }
         
         var lastReleaseUntil: Date?
@@ -64,6 +67,7 @@ class SentrySessionReplayTests: XCTestCase {
         var lastReplayRecording: SentryReplayRecording?
         var lastVideoUrl: URL?
         var lastReplayId: SentryId?
+        var currentScreen: String?
         
         func getSut(options: SentryReplayOptions = .init(sessionSampleRate: 0, errorSampleRate: 0) ) -> SentrySessionReplay {
             return SentrySessionReplay(replayOptions: options,
@@ -94,6 +98,10 @@ class SentrySessionReplayTests: XCTestCase {
         func breadcrumbsForSessionReplay() -> [Breadcrumb] {
             breadcrumbs ?? []
         }
+        
+        func currentScreenNameForSessionReplay() -> String? {
+            return currentScreen
+        }
     }
     
     override func setUp() {
@@ -104,14 +112,9 @@ class SentrySessionReplayTests: XCTestCase {
         super.tearDown()
         clearTestState()
     }
-    
-    private func startFixture() -> Fixture {
-        let fixture = Fixture()
-        return fixture
-    }
-    
+        
     func testDontSentReplay_NoFullSession() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         let sut = fixture.getSut()
         sut.start(rootView: fixture.rootView, fullSession: false)
         
@@ -124,7 +127,7 @@ class SentrySessionReplayTests: XCTestCase {
     }
     
     func testVideoSize() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         let options = SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1)
         let sut = fixture.getSut(options: options)
         let view = fixture.rootView
@@ -136,7 +139,7 @@ class SentrySessionReplayTests: XCTestCase {
     }
     
     func testSentReplay_FullSession() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         
         let sut = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
         sut.start(rootView: fixture.rootView, fullSession: true)
@@ -155,7 +158,7 @@ class SentrySessionReplayTests: XCTestCase {
             return
         }
         
-        XCTAssertEqual(videoArguments.duration, 5)
+        XCTAssertEqual(videoArguments.end, startEvent.addingTimeInterval(5))
         XCTAssertEqual(videoArguments.beginning, startEvent)
         XCTAssertEqual(videoArguments.outputFileURL, fixture.cacheFolder.appendingPathComponent("segments/0.mp4"))
         
@@ -164,8 +167,30 @@ class SentrySessionReplayTests: XCTestCase {
         assertFullSession(sut, expected: true)
     }
     
+    func testReplayScreenNames() throws {
+        let fixture = Fixture()
+        let sut = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
+        sut.start(rootView: fixture.rootView, fullSession: true)
+        
+        for i in 1...6 {
+            fixture.currentScreen = "Screen \(i)"
+            fixture.dateProvider.advance(by: 1)
+            Dynamic(sut).newFrame(nil)
+        }
+                
+        let urls = try XCTUnwrap(fixture.lastReplayEvent?.urls)
+        
+        guard urls.count == 6 else {
+        	XCTFail("Expected 6 screen names")
+        	return
+        }
+        XCTAssertEqual(urls[0], "Screen 1")
+        XCTAssertEqual(urls[1], "Screen 2")
+        XCTAssertEqual(urls[2], "Screen 3")
+    }
+    
     func testDontSentReplay_NotFullSession() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         let sut = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
         sut.start(rootView: fixture.rootView, fullSession: false)
         
@@ -184,7 +209,7 @@ class SentrySessionReplayTests: XCTestCase {
     }
     
     func testChangeReplayMode_forErrorEvent() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         let sut = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
         sut.start(rootView: fixture.rootView, fullSession: false)
         XCTAssertNil(fixture.lastReplayId)
@@ -197,7 +222,7 @@ class SentrySessionReplayTests: XCTestCase {
     }
     
     func testDontChangeReplayMode_forNonErrorEvent() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         let sut = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
         sut.start(rootView: fixture.rootView, fullSession: false)
         
@@ -210,7 +235,7 @@ class SentrySessionReplayTests: XCTestCase {
     
     @available(iOS 16.0, tvOS 16, *)
     func testChangeReplayMode_forHybridSDKEvent() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         let sut = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
         sut.start(rootView: fixture.rootView, fullSession: false)
 
@@ -222,7 +247,7 @@ class SentrySessionReplayTests: XCTestCase {
 
     @available(iOS 16.0, tvOS 16, *)
     func testSessionReplayMaximumDuration() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         let sut = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
         sut.start(rootView: fixture.rootView, fullSession: true)
         
@@ -238,7 +263,7 @@ class SentrySessionReplayTests: XCTestCase {
     
     @available(iOS 16.0, tvOS 16, *)
     func testDealloc_CallsStop() {
-        let fixture = startFixture()
+        let fixture = Fixture()
         func sutIsDeallocatedAfterCallingMe() {
             _ = fixture.getSut(options: SentryReplayOptions(sessionSampleRate: 1, errorSampleRate: 1))
         }